+++ src/screens/login/Modal.js
@@ -14,14 +14,40 @@ import React, {useState} from 'react';
 import styles from '../login/loginStyle';
 import LocalImages from '../../utils/localImages';
 import CustomButton from '../../components/button/customButton';
+import {useNavigation} from '@react-navigation/native';
 
-const Modal = () => {
+const initialState = {
+  users: [],
+  inputText: '',
+  confrimOtp: null,
+  switchFunction: false,
+  selection: {start: 0, end: 0},
+};
+
+function reducer(state, action) {
+  switch (action.type) {
+    case 'USERS':
+      return {...state, users: action.usersPayload};
+    case 'SELECTIONS':
+      return {...state, switchFunction: action.switchFunctionPayload};
+    case 'INPUT_TEXT':
+      return {...state, inputText: action.inputPayload};
+    case 'CONFIRM_OTP':
+      return {...state, confrimOtp: action.otpResponsePayload};
+    case 'SWITCH_FUNCTION':
+      return {...state};
+    default:
+      return null;
+  }
+}
+
+const Modal = ({callback}) => {
+  const [state, dispatch] = React.useReducer(reducer, initialState);
+  console.log('state', state);
   const keypadArray = [1, 2, 3, 4, 5, 6, 7, 8, 9, '+', 0, -1];
   const [inputText, setInputText] = useState('');
   const [selection, setSelection] = useState({start: 0, end: 0});
-  const [confrimOtp, setConfirmOtp] = useState(null);
-  const [count, setCount] = React.useState(false);
-  const [users, setUsers] = React.useState([]);
+  const navigation = useNavigation();
 
   const onPress = item => {
     if (item === -1 && selection !== null) {
@@ -58,9 +84,9 @@ const Modal = () => {
       signInWithPhoneNumberOtp(
         inputText,
         otpResponse => {
-          setCount(true);
-          setInputText('');
-          setConfirmOtp(otpResponse);
+          dispatch({type: 'SWITCH_FUNCTION', switchFunctionPayload: true});
+          dispatch({type: 'INPUT_TEXT', inputPayload: ''});
+          dispatch({type: 'CONFIRM_OTP', otpResponsePayload: otpResponse});
         },
         error => {
           console.log('error from the verification if catch', error);
@@ -71,17 +97,22 @@ const Modal = () => {
         inputText,
         confrimOtp,
         user => {
-          setUsers([...users, user]);
+          dispatch({type: 'SWITCH_FUNCTION', switchFunctionPayload: false});
+          dispatch({type: 'USERS', usersPayload: [...users, user.user._user]});
+          callback(true);
+          navigation.navigate('routes');
           console.log('user', user);
         },
         error => {
-          console.log('errro from the verfication else catch', error);
+          console.log('verification function catch', error);
         },
       );
     }
   };
 
-  const collection = () => {};
+  const collection = () => {
+    
+  };
 
   return (
     <View style={styles.modalMainView}>
